src/main/java/Converter/Converter.java
package Converter;

import java.util.Arrays;
import java.util.Comparator;
import java.util.List;
import java.util.stream.Collectors;

public class Converter {


    public Double fahrenheitToCelcius(double number) {
        double celcius = ((5 * (number - 32.0)) / 9.0);
        if (Double.isNaN(celcius)) {
            throw new NumberFormatException("Not a number");
        }
        return celcius;
    }

    public Double celciusToFahrenheit(double number) {
        double fahrenheit = ((number * 9) / 5) + 32;
        if (Double.isNaN(fahrenheit)) {
            throw new NumberFormatException("Not a number");
        }
        return fahrenheit;
    }


    enum RomanNumeral {
        I(1), IV(4), V(5), IX(9), X(10),
        XL(40), L(50), XC(90), C(100),
        CD(400), D(500), CM(900), M(1000);

        private int value;

        RomanNumeral(int value) {
            this.value = value;
        }

        public int getValue() {
            return value;
        }

        public static List<RomanNumeral> getReverseSortedValues() {
            return Arrays.stream(values())
                    .sorted(Comparator.comparing((RomanNumeral e) -> e.value).reversed())
                    .collect(Collectors.toList());
        }
    }

    public String convertArabicToRoman(int number) {
        if ((number <= 0) || (number > 4000)) {
            throw new IllegalArgumentException(number + " not in range");
        }

        List<RomanNumeral> romanNumerals = RomanNumeral.getReverseSortedValues();

        int i = 0;
        StringBuilder sb = new StringBuilder();

        while ((number > 0) && (i < romanNumerals.size())) {
            RomanNumeral currentSymbol = romanNumerals.get(i);
            if (currentSymbol.getValue() <= number) {
                sb.append(currentSymbol.name());
                number -= currentSymbol.getValue();
            } else {
                i++;
            }
        }

        return sb.toString();
    }
}
